+++ NEWS.md
@@ -1,3 +1,6 @@
+# ngme2 0.8.4 (2026-02-01)
+* Fix iid model using argument mesh instead of map
+
 # ngme2 0.8.3 (2025-12-10)
 * Update nu parameterization to be relative to nu_lower_bound (nu = nu_lower_bound + exp(B_nu * theta_nu))
 * Allow NULL specification for noise parameter in non-stationary case